+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 15.0.4
+## 15.0.4 (November 16, 2017)
 
 ### Enhancements
 - **Admin Changes**